- changing names of variables
