+++ e-store/e-store/ScenesScreen/CategoryItems/ViewController/CategoryItemsViewController.xib
@@ -9,7 +9,7 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="ItemViewController" customModule="e_store" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="CategoryItemsViewController" customModule="e_store" customModuleProvider="target">
             <connections>
                 <outlet property="itemsCollectionView" destination="JEO-cs-iOO" id="6lK-Ak-Skj"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
